Fix fetch_signal plugin and rename to is_loading_identifier
After talking with delaney I changed the behaviour of the plugin.

It now takes an identifier and adds it to a `Set` bound to a signal called `isLoading`,
if the identifier is present in the `Set`, then it means the query is fetching.

I changed the name of the plugin to reflect this and updated the example.