fix problem with sets in IsLoading backend plugin
We had basically a race condition where the signal store was being reset in
between the start and end of the fetch so the `s` was no longer pointing
to the right store. We also had an issue with sets having weird interactions
with signals and basically always setting the empty object instead of the
empty set, so we switched to arrays.